// src/Shared/Enums/Interview.enum.ts
export enum InterviewTypes {
  TECHNICAL = "Technical",
  HR = "HR",
  FINAL = "Final",
}

export enum InterviewStatus {
  SCHEDULED = "Scheduled",
  COMPLETED = "Completed",
  CANCELLED = "Cancelled",
  RESCHEDULED = "Rescheduled",
}

export enum InterviewNextStep {
  OFFERED = "Offered",
  REJECTED = "rejected",
  ANOTHERINTERVIEW = "Another Interview",
}